feat(viewer): convert app-wiring to a component-based model
