+++ libs/routers_viewer/src/components/mod.rs
@@ -1,0 +1,5 @@
+mod input;
+mod shell;
+
+pub use input::Input;
+pub use shell::Shell;
