+++ libs/routers_viewer/src/utils/mod.rs
@@ -1,5 +1,7 @@
-pub mod colour;
-pub mod component;
+mod colour;
+mod component;
+mod dimensions;
 
-pub use colour::ColourFactory;
+pub use colour::{BaseColour, ColourFactory, ColourScheme};
 pub use component::{Component, Context};
+pub use dimensions::{Layout, Regular, Size};
